# apps/backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kedai Bunda API",
        "description": "API untuk Sistem Kasir dan Manajemen Stok Kedai Bunda",
        "contact": {
            "email": "admin@kedaibunda.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/bahan-baku": {
            "get": {
                "tags": [
                    "Bahan Baku"
                ],
                "summary": "Mendapatkan daftar bahan baku",
                "operationId": "c2c1a78f757dffd53aeb1dbf356ba7ef",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data bahan baku",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Berhasil mengambil data bahan baku"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "Ayam Potong"
                                                    },
                                                    "satuan_dasar": {
                                                        "type": "string",
                                                        "example": "ekor"
                                                    },
                                                    "stok_tersedia": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 10.5
                                                    },
                                                    "harga_per_satuan": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 35000
                                                    },
                                                    "keterangan": {
                                                        "type": "string",
                                                        "example": "Ayam segar"
                                                    },
                                                    "aktif": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bahan Baku"
                ],
                "summary": "Menambah bahan baku baru",
                "operationId": "84118f8158e05bac6f69d4c7f60caa92",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "satuan_dasar",
                                    "stok_tersedia",
                                    "harga_per_satuan"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Ayam Potong"
                                    },
                                    "satuan_dasar": {
                                        "type": "string",
                                        "example": "ekor"
                                    },
                                    "stok_tersedia": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "harga_per_satuan": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 35000
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Ayam segar"
                                    },
                                    "aktif": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bahan baku berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sukses": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "pesan": {
                                            "type": "string",
                                            "example": "Bahan baku berhasil ditambahkan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}